+++ nalp/neurals/rnn.py
@@ -1,13 +1,11 @@
 import nalp.utils.logging as l
-import numpy as np
-import tensorflow as tf
 from nalp.core.neural import Neural
-from nalp.utils import math
+import tensorflow as tf
 
 logger = l.get_logger(__name__)
 
 
-class RNN(Neural):
+class RNN(tf.keras.Model):
     """A RNN class is the one in charge of Recurrent Neural Networks vanilla implementation.
     
     References:
@@ -15,7 +13,7 @@ class RNN(Neural):
 
     """
 
-    def __init__(self, vocab_size=1, hidden_size=2, learning_rate=0.001):
+    def __init__(self, vocab_size=1, hidden_size=2):
         """Initialization method.
 
         Args:
@@ -28,123 +26,20 @@ class RNN(Neural):
         logger.info('Overriding class: Neural -> RNN.')
 
         # Overrides its parent class with any custom arguments if needed
-        super(RNN, self).__init__()
-
-        # One for vocab size
-        self._vocab_size = vocab_size
-
-        # One for the amount of hidden neurons
-        self._hidden_size = hidden_size
-
-        # And the last for the learning rate
-        self._learning_rate = learning_rate
-
-        # Actually build the model
-        self._build()
-
-        logger.info('Class overrided.')
-
-    @property
-    def vocab_size(self):
-        """int: The size of the vocabulary.
-
-        """
-
-        return self._vocab_size
-
-    @property
-    def hidden_size(self):
-        """int: The amount of hidden neurons.
-
-        """
-
-        return self._hidden_size
-
-    @property
-    def learning_rate(self):
-        """float: A big or small addition on the optimizer steps.
-
-        """
-
-        return self._learning_rate
-
-    def _build(self):
-        """Main building method.
-
-        """
-
-        logger.info('Running private method: build().')
-
-        # Builds the model layers
-        self._build_layers()
-
-        # Builds the learning objects
-        self._build_learners()
-
-        # Builds the metrics
-        self._build_metrics()
-
-        logger.info('Model ready to be used.')
-
-    def _build_layers(self):
-        """Builds the model layers itself.
-
-        """
-
-        logger.debug(
-            f'Constructing model with shape: ({self.hidden_size}, {self.vocab_size}).')
+        super(RNN, self).__init__(name='rnn')
 
         # Creates a simple RNN cell
-        self.cell = tf.keras.layers.SimpleRNNCell(self.hidden_size)
+        self.cell = tf.keras.layers.SimpleRNNCell(hidden_size, name='rnn_cell')
 
         # Creates the RNN loop itself
-        self.rnn = tf.keras.layers.RNN(self.cell)
+        self.rnn = tf.keras.layers.RNN(self.cell, name='rnn_layer')
 
         # Creates the linear (Dense) layer
-        self.linear = tf.keras.layers.Dense(self.vocab_size)
+        self.linear = tf.keras.layers.Dense(vocab_size, name='dense')
 
         # And finally, a softmax activation for life's easing
-        self.softmax = tf.keras.layers.Softmax()
+        self.softmax = tf.keras.layers.Softmax(name='softmax')
 
-    def _build_learners(self):
-        """Builds all learning-related objects (i.e., loss and optimizer).
-
-        """
-
-        # Defining the loss function
-        self.loss = tf.losses.CategoricalCrossentropy()
-
-        logger.debug(f'Loss: {self.loss}.')
-
-        # Creates an optimizer object
-        self.optimizer = tf.optimizers.Adam(self.learning_rate)
-
-        logger.debug(
-            f'Optimizer: {self.optimizer} | Learning rate: {self.learning_rate}.')
-
-    def _build_metrics(self):
-        """Builds any desired metrics to be used with the model.
-
-        """
-
-        # Defining training accuracy metric
-        self.train_accuracy = tf.metrics.CategoricalAccuracy(
-            name='train_accuracy')
-
-        # Defining training loss metric
-        self.train_loss = tf.metrics.Mean(name='train_loss')
-
-        # Defining validation accuracy metric
-        self.val_accuracy = tf.metrics.CategoricalAccuracy(
-            name='val_accuracy')
-
-        # Defining validation loss metric
-        self.val_loss = tf.metrics.Mean(name='val_loss')
-
-        logger.debug(
-            f'Train Accuracy: {self.train_accuracy} | Train Loss: {self.train_loss} | Val Accuracy: {self.val_accuracy} | Val Loss: {self.val_loss}.')
-
-    @tf.function
     def call(self, x):
         """Method that holds vital information whenever this class is called.
 
@@ -166,56 +61,3 @@ class RNN(Neural):
         x = self.softmax(x)
 
         return x
-
-    def generate_text(self, dataset, start_text='', length=1, temperature=1.0):
-        """Generates a maximum length of new text based on the probability of next char
-        ocurring.
-
-        Args:
-            dataset (OneHot): A OneHot dataset object.
-            start_text (str): The initial text for generating new text.
-            length (int): Maximum amount of generated text.
-            temperature (float): The amount of diversity to include when sampling.
-
-        Returns:
-            A list containing a custom generated text (can be characters or words).
-
-        """
-
-        logger.info(f'Generating new text with length: {length}.')
-
-        # Defining variable to hold decoded generation
-        output_text = start_text
-
-        # Creating indexated tokens from starting text
-        tokens_idx = dataset.indexate_tokens(
-            list(start_text), dataset.vocab_index)
-
-        # Creating seed to be inputed to the predictor
-        seed = np.zeros(
-            (1, len(tokens_idx), dataset.vocab_size), dtype=np.float32)
-
-        # Iterate through maximum desired length
-        for _ in range(length):
-            # Iterate through all indexated tokens
-            for i, idx in enumerate(tokens_idx):
-                # Encodes each token into dataset's encoding
-                seed[0, i] = dataset.one_hot_encode(idx, dataset.vocab_size)
-
-            # Calculates the prediction
-            predict = self(seed).numpy()
-
-            # Chooses a index based on the predictions probability distribution
-            pred_idx = math.sample_from_multinomial(
-                predict[-1], temperature)
-
-            # Removing first indexated token
-            tokens_idx = np.delete(tokens_idx, 0, 0)
-
-            # Appending predicted index to the end of indexated tokens
-            tokens_idx = np.append(tokens_idx, pred_idx)
-
-            # Outputting generated characters to start text
-            output_text.append(dataset.index_vocab[pred_idx])
-
-        return output_text
